requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Configuration
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2

# PDF Processing
pymupdf>=1.23.0
img2pdf>=0.5.1

# Image Processing
pillow>=10.2.0

# HTTP client (for external APIs)
httpx>=0.26.0

# Email
resend>=0.7.0

# Stripe
stripe>=7.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
ruff>=0.1.0
